Changed the Structure.
Instead of getting the last point that was added to the DB, I now save the point n a Json file. Better when having a big amount of data.
When we add a new point,  we save the point to a json file. Then we read the point and update point before inserting into the DB

Co-Authored-By: Abdel Rahman Mallah <55826317+AbdoMallah@users.noreply.github.com>